satellite/metainfo: reduce begin segment response size
This is minimal change we can do to reduce begin segment response. We
are removing satellite signature from order limits embedded into
StreamID. We don't need this signature because whole StreamID is
signed.
We can optimize it a bit more but that would require more changes and
adjustments in RetryBeginSegmentPieces endpoint. That's the reason why
more complex test was added for RetryBeginSegmentPieces.

This change should make begin segment response ~10% smaller.

Change-Id: I8c389e4312cb2360d4056f8ce15ec17508527528