satellite/nodeselection: new weighted selector
With the current selectors it's not very easy to adjust load on
certain servers. We have different tracker based solutions
(like choiceofn), and they are good to exclude very bad nodes,
but not very flexible, and not so good to give significant more
load for a few servers.

This patch introduces a new selection algorithm, which is very
similar to random, but each node can have different (manually adjusted) weight.

It makes it possible to test load with manually adjusted balancing and
 evaluate the results.

Still choiceofn is probably best for all the time (like an autopilot), but for short period of time, the weighted selector can be a more flexible choice.

Change-Id: Idda8dd0309252b258ca3cf9745dc3af01248250f