scripts/release.sh: add back better ldflags
it has come to my attention that we do release.sh
for binaries other than storagenode and they may
not have been built on a tag with tag-release.sh
so they may not be set as release builds.

this changes release.sh to just always do a release
build because it's release.sh.

Change-Id: Ia3bad5d5fc4e6e079820f8b91774f07609701f53