web/satellite: drag and drop bug fix
Fix issue with drag and drop upload when duplicates are found. Chrome
clears the data transfer items after retrieving the File objects, so
when we retry uploading with the same DragEvent after the user agrees to
overwrite, the Files are no longer there. To solve this, get the Files
and store them in a ref, then pass the Files to be uploaded.

updates: https://github.com/storj/storj/issues/7172

Change-Id: Ifbf0607df2c5748cec31107efefdff79e7cef907