storagenode: full modular executor for 'storagenode exec'
Earlier I tried to create an alternative peer for storagenode run, but it turned to be easier just adding the `exec` and keep 'storagenode run' as is.

1. We don't need to touch existing 'storagenode run' (more safe)
2. we have more freedom to start our services

With the more freedom, we switched to a more simple approach:

 * the depednency hierarchy (mud.Ball) is assmbled before command is created
 * required components are defined by an environment variable

With this approach, we can filter out all the configuration which will be used (!) and bind them manually to parameters/config.

--> only the really used configuration will be printed out in --help

To achieve this, many helpers moved to 'shared/modular' to make them more re-usable.

Change-Id: I5d768fd32a29ada543ddcf852ce26c05c9871388