satellite/metainfo: Reduce uploads from 1001 to 10
Reduce the number of begin uploads from 1001 to 10 to reduce test time
and avoid that the test times out with the Spanner emulator.

I guess that 1001 was set to check that pagination works because Uplink
uses 1000 items per page. To trigger the pagination logic with 10 items
we set the limit through the context to 3 items per page.

Change-Id: If9f46c0eca0cbc459a141f8b08acd56338f089ca