satellite/metainfo: Adjust test & use proper testing variable
Adjust a test that used to fail with the Spanner emulator. Despite the
test didn't always fail in Spanner, we can reduce the number of object
from 10 to 5 go reduce the time to run and have less chances that the
test times out, which was the cause that we saw sometimes in the CI.

On the other hand, use the correct `testing.T` value in the sub-tests
and change a variable to a constant because it's used as that.

Change-Id: I40a7b0957d55ad09feb44df1b3b1877c83568b7b