storasgenode/piecestore: reduce dependency of endpoint
piecestore.Endpoint is a central component, depends on a lot of other components, which makes it harder to test it (and harder to disable unnecessary components).

In fact, Endpoint requires only limited functionality from other services (such as retain.Service, pieces.Deleter or pieces.TrashChore).

This patch cleans the dependency tree with using interfaces, to depend on just the required methods

Change-Id: I83e4e849b024e0b4d600f5db1323c56f3c474e85