storagenode/run: optional, alternative way to start modular storagenode
Some of the built-in functionality couldn't be disabled for modular startup without bigger refactor.

The plan is to create an experimental new CLI entrypoint which can be used for a pure modular experiment, without exposing it in release build.

Having all the logic in a non `main` package, will make it possible to enable optional plugins/modules similar to CoreDNS or Caddy.

(example: https://coredns.io/2017/07/25/compile-time-enabling-or-disabling-plugins/)

Change-Id: I745ba74b543b0bf3749fffb870865cbfcf007849