satellite/metainfo: Allow to set time func for endpoint
Add a field and a method the metainfo's Endpoint to allow to specify the
function that's called to get the current time for the rate limiter.

This allow to use a function in tests that returns always the same time
and prevent the rate limit tests to be flaky due to the difference of
execution times between different database systems.

Before this, Spanner emulator was not passing the test because it runs
quite slower than Postgres and Cockroach.

This commit also restore the rate limiter burst values set by the test
which were lowered to unsuccessfully mitigate the Spanner execution
slowness.

Change-Id: Iac27a6147663f1564866efd5977dac24694ebc33