shared/dbutil: consolidate db retrying logic
With distributed databases like CockroachDB and Google Cloud Spanner, it
is often necessary to abort transactions when they conflict with
concurrent modifications. The client code is expected to retry as
appropriate when this happens. Up to this point, we've mostly relied on
a cockroach-specific NeedsRetry function to make this determination.

However, Spanner indicates the need to retry in a different way. What's
more, it can also be correct to retry a transaction on PostgreSQL (it's
less likely to happen on a single-host database, but it can happen).

This commit consolidates transaction retry logic into one place which is
intended to handle all of our SQL-flavored database backends.

Change-Id: I525c624749a047b6ee758e85d8f1229d2820758e