satellite/satellitedb: cast at_rest_total to numeric before summing
The `::decimal` cast was added to the postgresql/cockroachdb version of
the query quite deliberately in commit 2ded64ba. Jeff thinks this was
probably due to the non-deterministic nature of summing over a floating
point field (floating point addition is not associative, so a somewhat
different sum could be calculated each time through). I noticed we did
not carry over this cast when porting to spanner, so I am implementing
it here.

Change-Id: Ibc5942fe859375297c2a99610687301d4a737cf0