go.mod,go.sum: Bump oidc package version
We initially used a lower version of the oidc package because our test
suite required the ability to build with go version 1.20 (due to legacy
storagenodes).
This script bumps the oidc package to the latest version, and uses an
override script to decrease the version of the package when building for
legacy storagenodes.

Change-Id: I9938e2b32b5267777946843efa72bda96a65b736