storagenode/pieces: pass batch size to expirations db when getting expired pieces
To ensure backward compatibility with nodes that were previously
using the pieceexpirationsdb before switching to the flat-files,
we try to also get expired pieces from the db as well.

Unfortunately we were discarding the limit and getting all
expired pieces from the db, but deleting up to the limit
so the nodes were left with a chunk of expired pieces that
have been deleted but still have data in the DB leading
to continuous "file not found" errors when the collector
tries to delete those pieces again.

Change-Id: Id36e167cca32730ad3e48b748aa5db6f19b56ed5