shared/dbutil/spannerutil: allow query of spanner schema
This is a complement to shared/dbutil/pgutil/query.go, which allows
inspection of the current postgresql/cockroach schema. Having this
implemented will allow us to test schema migrations for correctness
the same way we do in postgresql.

These changes will also allow a tool that performs comparison of two
databases and their contents.

Change-Id: Ie4f62a29c555cc1182d99ac97de6688756ee5054