shared/dbutil: don't sort PrimaryKey in dbschema.Schema
It is sometimes important to know the order in which columns appear in
the primary key of a database table. Records in the table are ordered
primarily according to the first column in the primary key, so someone
inspecting the table schema may need to know which one is first.

In dbschema.Table the PrimaryKey member is sorted lexically. This was
probably done to make it easier to compare tables where it is difficult
to determine the order of columns in the primary key, but we think it
was a mistake. Therefore, with this change we will no longer sort
PrimaryKey lexically and will instead leave the column names in the
order in which they appear in the primary key.

Change-Id: I9de814a000c98fe5760d5c87d5352bd21026fb68