storagenode/hashstore: incremental log rewriting
so imagine you have a node with like 50% ttl data and
50% live data spread across all the log files and all
the ttl data expires. now you have to rewrite all of
the log files and so if you have 100G of live data
you need to store 150GB of data peak or 50% overhead.

instead, if we bound the number of log files rewritten
per hashtable compaction, you only need O(1) overhead
at the cost of potentially writing the hashtable many
times, but that cost is smaller compared to the cost
of rewriting log files.

we do a middle ground where the number of log files
rewritten is unbounded but the live data written will
be less than the hashtable size, so we don't end up
in a situation where we write a 10G hashtbl, 1M log,
10G hashtbl, 1M log, etc.

Change-Id: I5d1323a259ebd410044a95d9bd79241235e12dc9