storagenode/hashstore: clump pieces by ttl
this makes it so that log file rewriting is much
cheaper under mixed load of ttl and non-ttl data
by making an effort to keep ttl pieces clumped
into log files. when we compact, if the whole log
file is dead, no writing of live pieces has to
occur and so we can just unlink the file and move
on.

the cost of this is that we may have more active
log files. i expect the number of active log files
to be around the max concurrency times the number
of active ttl days so this shouldn't be too bad.

Change-Id: I0e951be4127c32a42917d71d82926cd1146759e9