storagenode/hashstore: windows support
as far as i could find, windows would not allow me
to rename a file over an existing open file handle
which we rely on during hashtbl compaction to
atomically update the current hashtbl with the
next hashtbl. in order to support windows, the
hashtbl is now named hashtbl-<id> much like log
files and instead of renaming over, we just delete
the old one after renaming the new one to the
correct name. this means that the open code has to
find the largest committed hashtbl to open instead
of using a fixed name, but that's not too bad.

in order to support renaming an existing open file
on windows, though, we need to call the CreateFile
syscall ourselves with the FILE_SHARE_DELETE flag
set. unfortunately, the stdlib provides no way of
setting that flag.

Change-Id: I6fbbfdfe8ac145041dec152b597f6be454f24325