storagenode/hashstore: reduce max log size
this has a couple of benefits:

1. the maximum extra space used during compaction
is reduced to 0.75*1G instead of 0.75*10G. in the
worst case we may need to rewrite an entire log
file before we can unlink it regardless of hashtbl
size to ensure forward progress.

2. the variance on how much extra space is kept in
log files waiting to be reclaimed by compaction is
reduced. by having more files it's more likely
that some files will be above the rewrite
threshold. an easy way to see this is to consider
the extremes: if we had only 1 file that was just
below the threshold we would not rewrite it, but
if every piece was it's own file we would delete
every dead piece.

we will end up with more files total, but i think
this will still remain well below the threshold to
see wins.

Change-Id: Ide32bde7ac69626b8c5e12e6dad1e465b2843163