storagenode/hashstore: more robust startup sample
instead of the first 256 pages, just pick 256
random pages to make the estimation. this changes
fully linear reads to random reads but it'll only
be like a couple milliseconds either way.

this also loudly logs when we hit a table is full
error and attempts to redo the size estimations in
case something got wacky. this should never happen
so it doesn't need to be super efficient and can
help us track down what's going on if it does
happen. we also now track it in the stats so we
can monitor it.

it also removes the hashtbl page cache because i
doubt it provides too much benefit under mixed
workloads. instead, we have a cache per function
call because then they wont fight. this does mean
that repeated calls for finding a piece can't
reuse the page, but that was potentially the case
if the record was pushed out of it's home slot,
which isn't super rare.

Change-Id: I6ed5f5830d751d4c97cc7a038c0d4d7d93960b2b