storagenode/hashstore: randomize keys in table
if the incoming distribution of keys was ordered,
the hashtable would assign ordered slots to the
keys. this is bad because it causes the hashtbl to
have very long runs of filled slots, degrading
performance and causing problems with our old
estimation estimation algorithm that only looked
at the first pages. this change adds a call to
xxh3 to get the initial value for the slot before
shifting and masking, making the slot distribution
random even if the key distribution is ordered.

it is backwards compatible by using the header in
the hashtbl to store if key slot randomization is
enabled.

Change-Id: I8c48475d20ef371b601ead0b2dd3fd0fabc918b6