testsuite/rolling-upgrade: fix grep writing to closed pipe
When using:

  git tag -l --sort -version:refname | grep -v rc | head -1

The head may stop before grep has processed all the data, hence it will
end up writing to a closed pipe. Causing a "exit code 141" failure.

The fix is to consume the full output from grep to avoid the failure.

  { head -1; cat >/dev/null; }

An alternative would be to disable pipefail, but that may hide other
errors.

Change-Id: I6918eac14e8df39dc4d702531ed011a9c6354662