storagenode/hashstore: partition log directory
if we have a 24TB drive, with 1G files, that's 24k
files which is quite a bit. 256 is the closest power
of 2 such that it squared is greater than the largest
number of files we expect to reasonably have and gets
us to 64TB drives. it's also easy to change in the
future as the code will just load log files recursively
wherever they exist. that's also how it's backwards
compatible.

Change-Id: I8357c956355298a47f7d5501e3120500f02b5d59