storagenode/collector: reverse flat file store deletion order
We've killed many times the expiration jobs that were running the old
implementation, which left a lot of hour files that have already been
processed. Running the jobs again requires them to process these files
again, which significantly delays the deletion of the remaining expired
pieces.

By reversing the order, the job will priorities the expired pieces that
has not been deleted yet.

Change-Id: I95eaafe8ecc603fd939a34f91467b98766b715dd