storagenode/hashstore: clamp ttls before 1970 to 1970
before this, a ttl in the past would get clamped
up to the year 24937. the important property is
that the truncated timestamp is larger than the
input timestamp and for values earlier that 1970,
it's kinda weird to go up 20,000 years when we can
just go to 1970.

also export these functions because they may come
in handy for external tools that want to inspect
the store.

Change-Id: I65881e834cc4bdaa5c5c6fcc6b5632d7698fbce8