storagenode/pieces: fix pos tracking bug on blobs
the code to seek past the start of the header in
the read path would run every time the pos was
smaller than the header, but wouldn't update the
pos to the new value. this mean that if we had a
file that was very small or we did reads with a
small buffer we could end up doing the seek
multiple times and reading the same data over and
over.

Change-Id: I2807e9c21e100293a00ae26442a9d45ab5c3b801