storagenode/hashstore: fix deadlock with multiple revives
the deadlock requires 3 goroutines:

1. Compact holds activeMu.Lock and blocks
   acquiring rmu.Lock
2. reviveRecord holds reviveMu.Lock and blocks
   acquiring activeMu.RLock
3. reviveRecord holds rmu.RLock and blocks
    acquiring reviveMu.Lock

this is fixed by having reviveRecord always drop
rmu before attempting to acquire reviveMu so that
compaction can proceed. this does mean that we
miss a fast path where we can quickly update the
record in the hashtbl when compaction isn't
running without calling Create, but that's a very
small perf loss in a rare code path to avoid the
deadlock.

Change-Id: I20f6c28c5d46e0ef077b622d344118e0492e80bc