storagenode/hashstore: some Compact optimizations
make Range use larger read sizes. this increases
the throughput of the read portion of the
Compaction by about 30%.

cache monkit tasks and remove the context. this
reduces allocations and gets another 30% uplift in
throughput. it'd be nice if monkit had an api that
didn't allocate. future work for monkit i guess.

Change-Id: I5a296fa088dbf29e789423c09b14819dae8f460e