satellite: Add admin endpoint set status
Temporary, add a new endpoint to the old Admin API to be able to set set
a user status for not having to use SQL war crimes.

The commit also adds an entry in the UI, but it limits to only set the
status of the user to 'pending deletion'.

There are a few changes in the admin README due to that its table of
content was stale.

Change-Id: Ia4fc2023b7d96d7ad832d68ef2f7e6e08e26340d