storagenode/hashstore: background compact active first
i noticed that my node had not swapped which store
was active for many days and thought about it
before realizing the issue was that a change
introduced to avoid compacting a table if nothing
would change was causing the background compaction
to always try to compact the passive store, which
would succeed and do nothing, and so never
attempted to swap and compact the active store.
fix this by just trying the active one first. if
the active store needs compaction, then it will
swap first, which ensures that both stores get the
background treatment.

Change-Id: I7cf65884c255516c8c29509164665ef3bdc67f80