web/satellite: fix invite link behavior when sso is enabled
This fixes an issue where sign up and log in buttons will be disabled
when an email is provided in the url and SSO is enabled. This happens
because the form does not check whether SSO should be required of the
user or not.

Issue: https://github.com/storj/storj/issues/7286

Change-Id: I1bffe9823b9fe3d39743c5751b8e143aa1e97e0a