storagenode/hashstore: explicitly test some primitives
there were a couple of lines in some helper
primitives that didn't have coverage from the
higher level tests, so just do some basic testing
of them.

also i remembered that you can set a drpcsignal
with a nil error and then .Err() will return nil
but that would be bad, so change out to using a
signalError helper that always returns an error if
the signal is set.

Change-Id: I89fe96b06c7c24831b630f4e1f526c9460251447