storagenode/hashstore: Range should update stats
this was taken out before we had the exit early if
compaction would do nothing change, and so it
never was useful because whenever we did a Range
we would always throw away the hash table and
write a new one. but now that we have that
optimization, updating the stats can be useful.

Change-Id: I65e388f43fdb168d37a04a7375e9cb255e040d46