storagenode/hashstore: better background compact timing
we used to approximate an exponential distribution
of sleep durations by doing a poisson process
where every minute we checked if we should
compact. this didn't have an easy way of bounding
the maximum time, and waking up every minute isn't
very aesthetic. instead, we can use the power of
statistics to directly sample the exponential
distribution and sleep that long. this also means
we can bound the maximum sleep much more easily.

Change-Id: Ia2bca3939c84db1d72002f5d9ebfb0ed417e1ff7