satellite/metainfo: add migration mode flag
When this flag will be enabled all non read requests will be rejected
with ResourceExausted error. Potentially we are thinking about using
it during Spanner migration.

This change also adds new http endpoint to manipulate this flag
dynamically without restarting the service.

Implementation could be more generic for other flags we may need
in the future but for now, let's see if we will use it.

Change-Id: I141e6b240cc6a2d868e68ac62484915cdf4f12ae