storagenode/hashstore: speed up tests
this changes the starting hashtbl size and minimum
size to be 1MB instead of 8MB, which speeds up a
bunch of tests 8x. it should not affect anything
in production. reorganizes the constants and uses
them in appropriate places. reduces the amount of
records added in TestDB_BasicOperation because the
background compaction is tested for explicitly in
later tests. overall, the tests went from taking
36s on my machine to 2.5s, for a 14x speedup.

Change-Id: Ied2861cbfdf757fcf3a9a7d91a710dcfd692c03a