storagenode/collector: disable the migration backend in tests
This change disables the migration backend for collector tests to ensure
that pieces with TTL are not moved to the hashstore, allowing tests to
focus on the collection of expired pieces in the old store. This is
necessary because these tests are specifically designed to verify the
functionality of the old collector.

Change-Id: I638a248e4a38e02b4ce104531e0c5b507526e1c3