satellite/metabase: remove "less than new redundancy repair shares" check
The idea behind this check is that we didn't want to repair a segment in
a way that put it right back into the repair queue. However, with our
limited node pool in Select, this happens on a regular basis. Rather
than wasting the effort, let's recognize that it's better to add a _few_
new shares, and have the segment at least a little bit healthier.

Change-Id: Id3bbc61408bf0f8d67455be3e0cc93ba3cddaebc