satellite/{console,web}: improve cunoFS form submission
This change updates cunoFS form status on user settings on the back end
instead of the front end, reducing the number of requests for this form
to just one. This will eliminate the potential bug of the form
submission succeeding and the user setting update failing. This change
also adds a length validation for the comment field in the cunoFS form.

Issue: https://github.com/storj/storj-private/issues/1095

Change-Id: I66b79e537d25401c41090f75e549838c0361411b