satellite/metabase: make DeleteObjects use and return stream ID suffix
This change updates the DeleteObjects metabase method to accept the
stream ID suffixes of objects to deleted. In addition, this information
is returned for each deleted object.

This prepares the method for use by the metainfo RPC service, which
requires this information to construct object version IDs.

References storj/edge#515

Change-Id: I219ee379d967f32b9e0ee17f815263f5a5bc483f