satellite/{metabase,satellitedb}: fixes for spanner nightly
TestMigrateSpanner takes too long to execute. To fix, it needs to be
adjusted to only run the last few migrations, but still setup the tables
during database creation.

TestIterateLoopSegment/spanner_stale_reads returns a slightly different
error message on production spanner. Mainly differences in whitespace.
Add a possibility to write a regular expression in ErrText instead of a
strict string.

TestRepairQueue_Select_Concurrently seems to be highly contentions and
concurrency needs to be reduced. To fix the underlying issue the query
should be rewritten to be less contentions, somehow.

Change-Id: Ib22528045bc5683457b5299a09e4878541ca0c77