satellite/repair: counter to estimate effectiveness of a download shortcut
We always download full k pieces for each segments (at least) during segment repair, to reconstruct the original segment.

But during decomissioning large number of nodes (with GE), it might be enough to download unhealthy-retrievable pieces, instead of k pieces.

Which is also faster.

But is it really a faster and more effective way for repair?

This patch introduces two new counters to find it out...

Change-Id: I3c21900ded4975dc9483cd933a080d39ad4d082f