satellite/nodeselection: tag based NodeValue supports default value
In placement config we can use `tag:signer/name` syntax, when we need a value of the tag.

This patch introduces the support of `tag:signer/name?default` syntax as well, to make it possible to define a default value.

Can be useful for weighted selector, as weight can be defined as a tag, like :`tag:signer/name?1.0`

Change-Id: I0afcd22ecd4e7670a39262a91998999f0a376045