storagenode/diskmon: reduce the level of hardware monitoring
A long time ago, we used monkit-hw to collect lots of data about
storage nodes. After a long discussion, we removed monkit-hw
because we felt it might be too invasive to the privacy of our
storage node operators.

It appears that https://review.dev.storj.io/c/storj/storj/+/12860
brought it back.

This change is a bit more selective and keeps some hardware
monitoring, without reporting on things like how much
network traffic the node is experiencing, with or without
the storage node software, or all partitions that are in use
and disk space reports on them, whether or not the storage
node uses them.

Change-Id: I641703853ca7fa67a0d26ac7597af57f44e74b38