satellite/console: fixed CSRF cookie issue for separate frontend server
Separated the CSRF cookie from our regular CookieAuth scope so that it can be used independently of whether CookieAuth is defined or not.
The cookie is set in the appHandler on each app load or reload, and its value is sent to the client side via the frontend config endpoint to be used as a header value for specific endpoints.

Change-Id: I01426d17844fc1798588b031da0c9f75268e49e0