satellite/satellitedb: don't need to delete rollups in a tx anymore
As far as I can tell, this was only in a transaction so that we could
use the `THEN RETURN` part of the query. A bug in go-sql-spanner
required us to use an explicit ReadWriteTransaction any time we used
`THEN RETURN`. That bug should be fixed as of change-id
I617ddc17c358a028a9f1cde56929259fad58f212.

Change-Id: Idb2836ac0b10112e0e93b001229a4932e90ff12f