storagenode/hashstore: optimistic flock
if flock doesn't work on a test file, then don't
bother trying to flock on the real lock file. this
allows the code to work on setups that don't
support flock while still providing some measure
of safety for setups that do. this is the same
strategy as used for fallocate.

Change-Id: I9aad5a53da28b78475bdabf7ecd39f6630af2017