satellite/console: Attempt to fix/debug flaky TestDeleteAccount
Related to https://github.com/storj/storj/issues/7242

After a lot of debugging, I was still unable to find an obvious issue
with the test which could explain its flakiness.
This commit includes one change to insert egress an hour after the
beginning of the billing window, rather than exactly at the beginning of
the billing window, to avoid a possible edge case there.
But I am not convinced this actually explains the flakiness, as I have
not been able to reproduce this edge case locally, and it seems to work,
even with egress inserted exactly at the start of the billing window.

Because I am not satisfied with the bug fix, I also introduced some
conditional debug logging which should help if/when the test is flaky in
the future.

Change-Id: Ib9b4f54e09dfcda9087dc46f9c4def0246aab816