satellite/nodeselection/failuretracker: use a stochastic rolloff from the failure tracker
right now, the failure tracker is somewhat of a conveyor belt where node
success counts are kept on 30s slots over a 2 minute window. when any given
30s elapses, all at once, all of the nodes that were bad 2 minutes ago
roll off the failure tracker conveyor belt and are readded to the
available pool of nodes together immediately. this stinks and we should
probably be more gradual about letting nodes come off the failure tracker.

to do that, this change introduces a stochastic property, where a slot
on the conveyor belt is not zeroed with some configurable probability.
over time a good node will end up moving off, but not alongside all
other nodes at the same time.

Change-Id: If3dc07ba899bdcd4d9cf78f46322b8dfa9ad697c