storagenode/hashstore: optimize rewriting records
by using a specially crafted io.Reader that the
stdlib knows how to special case, we get to do
the copy_file_range syscall which avoids copying
the data into userspace and filesystems may even
implement it better than copying the data. this
should improve compaction speeds for when we need
to rewrite lots of log files. on my machine, the
throughput is increased 20x.

    old.txt    │              new.txt               │
    sec/op     │   sec/op     vs base               │
 113.333µ ± 5%   5.333µ ± 2%  -95.29% (p=0.000 n=8)

    old.txt    │                new.txt                 │
      B/s      │      B/s       vs base                 │
 1.683Gi ± 12%   35.774Gi ± 2%  +2025.66% (p=0.000 n=8)

Change-Id: If2031185705b2f92199aa2af66e1dc33a68d8e3c