storagenode/hashstore: stochastic log rewriting
before this change, the only time a log would be
rewritten is when it had enough dead data to push
it below the AliveFraction (which defaulted to
0.75). this had a couple of implications:

    1. often, many logs would fall under this
    threshold at once, so we would go from doing
    no log rewriting to all of the log rewriting
    which is not good.
    2. since the file looked like [AAAD], where A
    is alive data and D is dead data, we would
    have to rewrite AAA to reclaim D, meaning we
    would write 3x as much data as we reclaimed.

to avoid problem 2 we've tried tuning this number
to be lower (like 0.01 or 0.25, etc) but this just
delays the problem: eventually we will have to
rewrite a ton of data all at once. this change
avoids problem 1 by rewriting logs stochastically
when they are above the AliveFraction with a
probability chosen to make the cost of rewriting
amortized constant.

to derive this probability, let's start with
some assumptions and design constraints:

    - alive is a percent in [0, 1] describing
      what percent of the file is alive.
    - we want cost(alive) * prob(alive) = C
      so that the cost multiplied by the
      probability of choosing it is constant
      so that we have constant cost over time.
    - cost(1) = infinity so that rewriting
      a log to reclaim 0 bytes is "expensive"
    - cost(0) = 0 so that writing 0 bytes
      to reclaim all the data is "cheap"

we can then find a cost function that makes
sense. a very simple one is:

    cost(alive) = alive / (1 - alive)

intuitively, this cost function answers the
question "how much alive data do we rewrite per
byte of dead data?" so that cost(0.5) is 1 and
cost(0.75) is 3 like above. solving for the
probability function, we get

    cost(alive) * prob(alive) = C
    alive / (1 - alive) * prob(alice) = C
    prob(alive) = C * (1 - alive) / alive

the remaining question is how to choose the
constant C. first, note that prob as defined is
monotonically decreasing with alive. therefore
picking C can be thought of as picking a threshold
so that any value less alive than some critical
value is rewritten, where a larger C value
increases the probability and lowers the
threshold.

we already have such a threshold, which is the
AliveFraction. we can take that and solve for a C
by solving (framed in terms of alive):

    1 = C * (1 - AliveFraction) / AliveFraction
    C = AliveFraction / (1 - AliveFraction)

and so our function is defined and constants
picked.

as a minor implementation note, because the
probability function is decreasing with alive (or,
increasing with dead), we don't have to bound it
by 1 to make it a "true" probability because we
choose by rolling a number in [0, 1) and comparing
it to the probability, so values greater than 1
are treated as if they are 1.

Change-Id: I2ab0f44b72b5124001812731afc49043bd05348a