satellite/repair: fix retry logic in AdminFetchPieces
During repair, we try first to connect to each node using its last IP
address and port. If we fail to connect that way, we try again using
the hostname.

But in AdminFetchPieces, we checked for a connection error incorrectly,
and ended up retrying the connection for (basically) _any_ error. So a
"piece not found" error would cause another try, which would lead to a
"used serial already exists in store" error the second time.

This bug only affects AdminFetchPieces, which is only used by incidental
management tools.

Change-Id: I8b4612079754c2708546f34a0b3688b8012552e0