satellite/durability: count generic health matrix
Durability Ranged Loop observer is created to simulate the event when some of the node groups become onffline at the same time. (For example: datacenter outage).

It calculates it with loosing 0,1 or 2 groups from each segments.

0 shows the current state (0 offline groups in addition to the current ones).

Most of the time we use this zero report, but during GE, it's quite confusing, as all nodes (and stored pieces) are counted as unhealthy.

Probably it's better to differentiate between un-healthy and retrievable pieces.

Pieces on GE nodes are retrievable, but not healthy.

Change-Id: Ie0eeb6ed96bbd3c411bdd7d53c2ecf3cc900783c